+++ fdm-core/src/cultivation.test.ts
@@ -1,7 +1,7 @@
-import { describe, expect, it, afterAll, beforeEach, beforeAll } from 'vitest'
+import { describe, expect, it, afterAll, beforeEach} from 'vitest'
 import { createFdmServer, migrateFdmServer } from './fdm-server'
 import { type FdmServerType } from './fdm-server.d'
-import { addCultivationToCatalogue, getCultivationsFromCatalogue, addCultivation, removeCultivation, getCultivation, getCultivations, getCultivationPlan } from './cultivation'
+import { addCultivationToCatalogue, getCultivationsFromCatalogue, addCultivation, removeCultivation, getCultivation, getCultivations, getCultivationPlan, updateCultivation } from './cultivation'
 import { addFarm } from './farm'
 import { addField } from './field'
 import { nanoid } from 'nanoid'
@@ -83,7 +83,7 @@ describe('Cultivation Data Model', () => {
 
         it('should throw an error when adding a cultivation with an invalid catalogue ID', async () => {
             const invalid_b_lu_catalogue = 'invalid-catalogue-id';
-            const b_sowing_date = '2024-01-01';
+            const b_sowing_date = new Date('2024-01-01');
 
             await expect(
                 addCultivation(fdm, invalid_b_lu_catalogue, b_id, b_sowing_date)
@@ -100,7 +100,7 @@ describe('Cultivation Data Model', () => {
                 b_lu_hcat3_name: 'test-hcat3-name'
             })
 
-            const b_sowing_date = '2024-01-01'
+            const b_sowing_date = new Date('2024-01-01')
             const b_lu = await addCultivation(fdm, b_lu_catalogue, b_id, b_sowing_date)
             expect(b_lu).toBeDefined()
 
@@ -120,7 +120,7 @@ describe('Cultivation Data Model', () => {
                 b_lu_hcat3_name: 'test-hcat3-name'
             });
 
-            const b_sowing_date = '2024-01-01';
+            const b_sowing_date = new Date('2024-01-01');
             await addCultivation(fdm, b_lu_catalogue, b_id, b_sowing_date);
 
             // Attempt to add the same cultivation again
@@ -139,29 +139,12 @@ describe('Cultivation Data Model', () => {
                 b_lu_hcat3_name: 'test-hcat3-name'
             });
 
-            const b_sowing_date = '2024-01-01';
+            const b_sowing_date = new Date('2024-01-01');
             const invalid_b_id = 'invalid-field-id';
 
             await expect(
                 addCultivation(fdm, b_lu_catalogue, invalid_b_id, b_sowing_date)
             ).rejects.toThrow('Field does not exist');
-        });
-
-        it('should throw an error when adding a cultivation with an invalid sowing date', async () => {
-            await addCultivationToCatalogue(fdm, {
-                b_lu_catalogue,
-                b_lu_source: 'test-source',
-                b_lu_name: 'test-name',
-                b_lu_name_en: 'test-name-en',
-                b_lu_hcat3: 'test-hcat3',
-                b_lu_hcat3_name: 'test-hcat3-name'
-            });
-
-            const invalid_b_sowing_date = 'invalid-date';
-
-            await expect(
-                addCultivation(fdm, b_lu_catalogue, b_id, invalid_b_sowing_date)
-            ).rejects.toThrow('Invalid sowing date');
         });
 
         it('should get cultivations by field ID', async () => {
@@ -174,8 +157,8 @@ describe('Cultivation Data Model', () => {
                 b_lu_hcat3_name: 'test-hcat3-name'
             })
 
-            await addCultivation(fdm, b_lu_catalogue, b_id, '2024-01-01')
-            await addCultivation(fdm, b_lu_catalogue, b_id, '2024-03-01')
+            await addCultivation(fdm, b_lu_catalogue, b_id, new Date('2024-01-01'))
+            await addCultivation(fdm, b_lu_catalogue, b_id, new Date('2024-03-01'))
 
             const cultivations = await getCultivations(fdm, b_id)
             expect(cultivations.length).toBe(2)
@@ -192,7 +175,7 @@ describe('Cultivation Data Model', () => {
                 b_lu_hcat3_name: 'test-hcat3-name'
             })
 
-            const b_sowing_date = '2024-01-01'
+            const b_sowing_date = new Date('2024-01-01')
             const b_lu = await addCultivation(fdm, b_lu_catalogue, b_id, b_sowing_date)
 
 
@@ -200,6 +183,71 @@ describe('Cultivation Data Model', () => {
 
             await expect(getCultivation(fdm, b_lu)).rejects.toThrowError('Cultivation does not exist')
         })
+
+        it('should update an existing cultivation', async () => {
+            await addCultivationToCatalogue(fdm, {
+                b_lu_catalogue,
+                b_lu_source: 'test-source',
+                b_lu_name: 'test-name',
+                b_lu_name_en: 'test-name-en',
+                b_lu_hcat3: 'test-hcat3',
+                b_lu_hcat3_name: 'test-hcat3-name'
+            })
+
+            const originalSowingDate = new Date('2024-01-01');
+            const b_lu = await addCultivation(fdm, b_lu_catalogue, b_id, originalSowingDate)
+
+            const newSowingDate = new Date('2024-02-01');
+            const newCatalogueId = nanoid();
+
+            // Add the new cultivation to the catalogue first
+            await addCultivationToCatalogue(fdm, {
+                b_lu_catalogue: newCatalogueId,
+                b_lu_source: 'new-source',
+                b_lu_name: 'new-name',
+                b_lu_name_en: 'new-name-en',
+                b_lu_hcat3: 'new-hcat3',
+                b_lu_hcat3_name: 'new-hcat3-name'
+            });
+
+
+            await updateCultivation(fdm, b_lu, newCatalogueId, newSowingDate);
+
+            const updatedCultivation = await getCultivation(fdm, b_lu);
+            expect(updatedCultivation.b_sowing_date).toEqual(newSowingDate);
+            expect(updatedCultivation.b_lu_catalogue).toEqual(newCatalogueId);
+
+
+        })
+
+        it('should throw an error when updating a non-existent cultivation', async () => {
+            const nonExistentBlu = nanoid();
+            await expect(updateCultivation(fdm, nonExistentBlu, b_lu_catalogue, new Date())).rejects.toThrowError('Cultivation does not exist');
+        });
+
+        it('should throw an error when updating with invalid catalogue id', async () => {
+
+            // Add the new cultivation to the catalogue first
+            const b_lu_catalogue = nanoid();
+
+            await addCultivationToCatalogue(fdm, {
+                b_lu_catalogue: b_lu_catalogue,
+                b_lu_source: 'new-source',
+                b_lu_name: 'new-name',
+                b_lu_name_en: 'new-name-en',
+                b_lu_hcat3: 'new-hcat3',
+                b_lu_hcat3_name: 'new-hcat3-name'
+            });
+
+
+            const originalSowingDate = new Date('2024-01-01');
+            const b_lu = await addCultivation(fdm, b_lu_catalogue, b_id, originalSowingDate)
+
+
+            const nonExistentCatalogueId = nanoid();
+
+            await expect(updateCultivation(fdm, b_lu, nonExistentCatalogueId, new Date())).rejects.toThrowError('Cultivation in catalogue does not exist');
+        });
 
 
     })
@@ -234,7 +282,7 @@ describe('Cultivation Data Model', () => {
                 b_lu_hcat3_name: "test"
             })
 
-            b_lu = await addCultivation(fdm, b_lu_catalogue, b_id, '2024-03-01')
+            b_lu = await addCultivation(fdm, b_lu_catalogue, b_id, new Date('2024-03-01'))
 
             // Add fertilizer to catalogue (needed for fertilizer application)
             const p_id_catalogue = nanoid();